Refactor and add dialog
